requirements.txt
openai
pandas
numpy
requests
PyYAML
matplotlib
seaborn
tqdm
scikit-learn
